Split resolver metadata requirement helper tests
